fix: Ensure tests can succeed without LocalStack running
The tests work fine if you have LocalStack running as well but they break if you do not. This commit fixes this by disabling Spring Cloud AWS SQS + SNS for the specific tests using an application-test.yml. Additionally, a MockitoBean has been added for SnsTemplate so that the producers can still get the necessary bean.