refac: Add DeduplicationService for all processed events (#5)
Add `DeduplicationService` which allows us to easily determine if an event has already been processed and record it as processed as well. Update `OrderService` to also use deduplication. The methods in `OrderService` are already idempotent because multiple consumed identical events would simply set the order status to what it has already been set to. However, I have added deduplication for this service as well to get into the habit of using deduplication. This is a good showcase of how you can implement deduplication in a project.