refac: Remove id from OutboxEntity
OutboxEntity has previously had both an id and eventId. This might be unnecessary as the outbox entity simply relates to an event that has not been published yet. So we can just use the eventId as the primary key for the outbox table. This commit also updates one of the Liquibase yaml changelogs to not create a column for id and only create a column for event_id as the primary key. Generally, you should never update an existing Liquibase yaml changelog. Considering that we do not have anything running in production or some live service running this, we can do it in this case. However, keep in mind that you should never update an existing Liquibase yaml changelog if you have some live instance running somewhere.